footbal_engine v1.0.0 FootbalEngine.Populator.Server

Server for the Cache. Tries to populate it with data and if it gets anything other than a complete success for the indexation, it will keep trying to repopulate the memory tables.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(path)

start_link(String.t()) :: GenServer.on_start()