footbal_engine v0.1.0 FootbalEngine.Populator.Storage

Stores the given data in memory, indexing it as necessary along the way.

Can return a complete success (if all the data was valid and inserted into memory), a partial success (if some data was faulty) or and error (depending on the reason it failed).

No data is also treated as an error case because an empty memory DB is useless and is therefore likely to be result of a human error.

Functions

persist(arg, deps)

persist({[map()], [any()]}, keyword()) ::
  {:ok, :indexation_successful}
  | {:ok, :partial_indexation_successful, [any()]}
  | {:error, :no_valid_data_to_save, [any()]}